requirements.txt
pandas
sentence-transformers
faiss-cpu
streamlit
ollama
transformers
scikit-learn
